+++ R/preflib.R
@@ -27,7 +27,7 @@
 #' giving the frequency of the ranking in that row, and remaining columns
 #' \code{Rank 1} \ldots \code{Rank p} giving the items ranked from first to
 #' last place in that ranking. Ties are represented by vector elements in list
-#' columns. The data frame has an attribute \code{"item"} giving the labels
+#' columns. The data frame has an attribute \code{"items"} giving the labels
 #' corresponding to each item number.
 #' @references
 #' Mattei, N. and Walsh, T. (2013) PrefLib: A Library of Preference Data.
@@ -45,7 +45,7 @@
 #' # strict complete orderings of four films on Netflix
 #' netflix <- read.soc(file.path(preflib, "netflix/ED-00004-00000101.soc"))
 #' head(netflix)
-#' attr(netflix, "item")
+#' attr(netflix, "items")
 #'
 #' head(as.rankings(netflix))
 #'
@@ -120,5 +120,5 @@ read.toi <- function(file){
 #' @export
 as.rankings.preflib <- function(x, verbose = TRUE, ...){
     as.rankings.matrix(as.matrix(x[, -1]), freq = x[, 1], input = "ordering",
-                       labels = attr(x, "item"), verbose = verbose, ...)
+                       items = attr(x, "items"), verbose = verbose, ...)
 }